// streampark-console/streampark-console-webapp-v2/src/enums/flinkEnum.ts
export enum ReleaseStateEnum {
    /** release failed */
    FAILED = -1,
    /** release done */
    DONE = 0,
    /** need release after modify task */
    NEED_RELEASE = 1,
    /** releasing */
    RELEASING = 2,
    /** release complete, need restart */
    NEED_RESTART = 3,
    /** need rollback */
    NEED_ROLLBACK = 4,
    /** project has changed, need to check the jar whether to be re-selected */
    NEED_CHECK = 5,
    /** revoked  */
    REVOKED = 10,
}
// ...
export enum PipelineStepEnum {
    UNKNOWN = 0,
    WAITING = 1,
    RUNNING = 2,
    SUCCESS = 3,
    FAILURE = 4,
    SKIPPED = 5,
}